Abstract
Disasters are what become of hazards that victimize people. To prevent disasters from occurring, every member of a society needs to be aware and prepared against them. This study aims to measure the awareness and preparedness levels for different segments of community separated by various DESG factors, and how open every segment is to be improved. The study was conducted by focusing mainly on earthquakes, floods and the recent pandemic disaster. First, an online survey was conducted with the aim of reaching every DESG group which was almost perfectly achieved, with a side-effect of possible biases. Later on, the data consisting of 639 subjects was examined using MS Excel, R and Python. After cleaning the data, measures for disaster awareness and preparedness were calculated using simple functions in R and MS Excel respectively. The DESG information was then visualized, and summary statistics were shown. Subsequently, relevant statistical methods such as Kruskal-Wallis tests for relationships between variables, Shapiro-Wilk tests to measure whether the variables were parametric, and Chi-Squared tests along with Cramer’s V correlation tests to look for relationships between categorical variables were used. Satisfying evidence on relationships between disaster awareness and age, region of accommodation; and satisfying evidence on relationships between disaster preparedness and age, income levels of participants were found. It was also observed that females are more willing to get educated on disaster awareness and preparedness.